refactor: move git panel into page slice
